+++ services/dli/v1/model/model_list_elastic_resource_pool_scale_records_response.go
@@ -10,7 +10,7 @@ import (
 type ListElasticResourcePoolScaleRecordsResponse struct {
 
 	// 返回数组长度
-	Count *int32 `json:"count,omitempty"`
+	Count *int64 `json:"count,omitempty"`
 
 	// 数组中返回的数据
 	Items *[]ElasticResourcePoolScaleRecord `json:"items,omitempty"`
